:recycle: Componentize Database page to facilitate changes
