+++ src/app/database/components/array-container.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+import ObjectContainer from './object-container';
+
+const ArrayContainer = styled(ObjectContainer)`
+    flex-direction: row;
+    gap: 1rem;
+`;
+
+export default ArrayContainer;
