+++ src/app/database/components/object-container.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const ObjectContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    border: 0.25rem solid black;
+    margin: 1rem;
+    border-radius: 1rem;
+    overflow: clip;
+    width: fit-content;
+`;
+
+export default ObjectContainer;
