+++ src/components/fab.tsx
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+const FAB = styled.button`
+    position: fixed;
+    bottom: 1rem;
+    right: 1rem;
+    padding: 1rem;
+    border-radius: 50%;
+    background-color: #000;
+    color: #fff;
+    border: none;
+    cursor: pointer;
+    font-size: 2rem;
+    font-weight: bold;
+    box-shadow: 0 0 1rem rgba(0, 0, 0, 0.5);
+`;
+
+export default FAB;
